+++ pkg/ingester/limiter.go
@@ -11,8 +11,6 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/grafana/mimir/pkg/util"
-	"github.com/grafana/mimir/pkg/util/globalerror"
-	"github.com/grafana/mimir/pkg/util/log"
 	util_math "github.com/grafana/mimir/pkg/util/math"
 	"github.com/grafana/mimir/pkg/util/validation"
 )
@@ -38,7 +36,7 @@ type Limiter struct {
 	ring                 RingCount
 	replicationFactor    int
 	zoneAwarenessEnabled bool
-	sampler              *log.Sampler
+	samplers             ingesterErrSamplers
 }
 
 // NewLimiter makes a new in-memory series limiter
@@ -54,7 +52,7 @@ func NewLimiter(
 		ring:                 ring,
 		replicationFactor:    replicationFactor,
 		zoneAwarenessEnabled: zoneAwarenessEnabled,
-		sampler:              log.NewSampler(errorSampleRate),
+		samplers:             newIngesterErrSamplers(errorSampleRate),
 	}
 }